# refactored & simplified Settings.hpp and pmtv format