--- a/core/include/gnuradio-4.0/Tag.hpp
+++ b/core/include/gnuradio-4.0/Tag.hpp
@@ -5,8 +5,7 @@
 
 #include <pmtv/pmt.hpp>
 
-#include <pmtv/format.hpp>
-
+#include <gnuradio-4.0/meta/formatter.hpp>
 #include <gnuradio-4.0/meta/utils.hpp>
 
 #include "reflection.hpp"