## Details:

* code and templating structure (via and in response to -ftime-trace analysis)
* replaced std::ranges::find(..) by more compile-time friendly version
* replaced std::sort by binary-search+merge-sort
* added non-inline to methods to avoid unnecessary compile-time optimisations
* optimised custom pmtv formatted by reducing the `std::visit(...)` pattern (not yet fully eliminated.

Compile-time differences (~4% ... hoped to be more, but it's a start).

**before:**
```text
real    20m21.229s
user    89m12.953s
sys     3m8.205s
```

**this PR:**
```text
real    19m33.855s
user    85m44.586s
sys     3m35.693s
```

ClangAnalyser profile with this PR (based on Clang's `-ftime-trace` feature) in the `./build` directory:
```bash
ClangBuildAnalyzer --all . gr4_all_before.bin; ClangBuildAnalyzer --analyze gr4_all_before.bin > gr4_all_before.txt
```
**before:**  [gr4_all_before.txt](https://github.com/user-attachments/files/16884919/gr4_all_before.txt)
```text
Analyzing build trace from 'gr4_all_before.bin'...
**** Time summary:
Compilation (184 times):
  Parsing (frontend):         2265.8 s
  Codegen & opts (backend):   8065.3 s
  [..]
```
**after:** [gr4_all_after.txt](https://github.com/user-attachments/files/16875660/gr4_all_after.txt)
```text
Analyzing build trace from 'gr4_all_after.bin'...
**** Time summary:
Compilation (184 times):
  Parsing (frontend):         2121.1 s
  Codegen & opts (backend):   4473.3 s
  [..]
```